# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * RescueNet AI - Disaster Crisis Triage & Dispatch System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Incident {
  id: number;
  rawText: string;
  /** @nullable */
  originalLanguage?: string | null;
  /** @nullable */
  translatedText?: string | null;
  /** medical | rescue | food | water | shelter | other */
  needType: string;
  /** critical | high | medium | low */
  urgency: string;
  /** @nullable */
  peopleAffected?: number | null;
  /** @nullable */
  summary?: string | null;
  /** 0-100 priority score */
  priorityScore: number;
  /**
     * AI confidence percentage 0-100
     * @nullable
     */
  aiConfidence?: number | null;
  /** @nullable */
  aiExplanation?: string | null;
  /**
     * JSON array of reasons as string
     * @nullable
     */
  aiReasoningFactors?: string | null;
  /** pending | ai_processing | assigned | dispatched | en_route | in_progress | resolved | closed */
  status: string;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  contactInfo?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  assignedResourceId?: number | null;
  /** @nullable */
  assignedResourceName?: string | null;
  /** @nullable */
  dispatchJustification?: string | null;
  /** @nullable */
  disasterType?: string | null;
  /** @nullable */
  incidentCategory?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface IncidentInput {
  /** @minLength 1 */
  rawText: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  /** @nullable */
  contactInfo?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface IncidentPatch {
  /** @nullable */
  status?: string | null;
  /** @nullable */
  assignedResourceId?: number | null;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  /** @nullable */
  address?: string | null;
}

export interface AssignmentInput {
  resourceId: number;
  /** @nullable */
  justification?: string | null;
}

export interface Resource {
  id: number;
  name: string;
  /** medical | rescue | food | shelter | fire | police | boat */
  type: string;
  /** available | busy | en_route | offline */
  status: string;
  /** @nullable */
  capacity?: number | null;
  lat: number;
  lng: number;
  isAvailable: boolean;
  /** @nullable */
  assignedIncidentId?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface ResourceInput {
  name: string;
  type: string;
  /** @nullable */
  capacity?: number | null;
  lat: number;
  lng: number;
}

export interface ResourcePatch {
  /** @nullable */
  status?: string | null;
  /** @nullable */
  isAvailable?: boolean | null;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  /** @nullable */
  assignedIncidentId?: number | null;
}

export interface ResourceRecommendation {
  resource: Resource;
  /** Distance in km */
  distance: number;
  reason: string;
  /** e.g. '4 minutes' */
  eta: string;
  /** 0-100 */
  confidence: number;
}

export interface TimelineEntry {
  id: number;
  incidentId: number;
  action: string;
  /** @nullable */
  details?: string | null;
  timestamp: string;
}

export interface DashboardStats {
  totalIncidents: number;
  criticalIncidents: number;
  pendingIncidents: number;
  resolvedIncidents: number;
  availableResources: number;
  resourcesInUse: number;
  avgResponseTimeMinutes: number;
  avgResolutionTimeMinutes: number;
  resourceUtilizationPercent: number;
}

export interface ChartDataPoint {
  name: string;
  value: number;
}

export interface TrendDataPoint {
  date: string;
  count: number;
}

export interface AnalyticsData {
  urgencyDistribution: ChartDataPoint[];
  typeDistribution: ChartDataPoint[];
  statusDistribution: ChartDataPoint[];
  dailyTrend: TrendDataPoint[];
  resolutionRate: number;
}

export interface ActivityItem {
  id: number;
  action: string;
  /** @nullable */
  resourceName?: string | null;
  /** @nullable */
  incidentId?: number | null;
  /** @nullable */
  incidentSummary?: string | null;
  /** @nullable */
  urgency?: string | null;
  timestamp: string;
}

export type ListIncidentsParams = {
/**
 * @nullable
 */
status?: string | null;
/**
 * @nullable
 */
urgency?: string | null;
/**
 * @nullable
 */
needType?: string | null;
/**
 * @nullable
 */
sortBy?: string | null;
/**
 * @nullable
 */
limit?: number | null;
};

export type ListResourcesParams = {
/**
 * @nullable
 */
type?: string | null;
/**
 * @nullable
 */
isAvailable?: boolean | null;
};

export type ListActivityParams = {
/**
 * @nullable
 */
limit?: number | null;
};